drm/gma500: add atomic support
Imported from https://git.tizen.org/cgit/kernel/kernel-mfld-blackbay/

This is currently not functional and based off of an older version of
the gma500 driver.

I'm looking to see if this would be a decent starting point for adding
gma500 atomic support to mainline and how best to approach that.

Signed-off-by: James Hilliard <james.hilliard1@gmail.com>